First implementation refactor plot.py
