+++ src/CSET/operators/__init__.py
@@ -30,6 +30,7 @@ from CSET.operators import (
     aggregate,
     aviation,
     collapse,
+    colormaps,
     constraints,
     convection,
     ensembles,
@@ -43,6 +44,7 @@ from CSET.operators import (
     pressure,
     read,
     regrid,
+    spectra,
     temperature,
     transect,
     wind,
@@ -55,6 +57,7 @@ __all__ = [
     "aggregate",
     "aviation",
     "collapse",
+    "colormaps",
     "constraints",
     "convection",
     "ensembles",
@@ -70,6 +73,7 @@ __all__ = [
     "pressure",
     "read",
     "regrid",
+    "spectra",
     "temperature",
     "transect",
     "wind",
