+++ tests/operators/test_plot.py
@@ -21,226 +21,10 @@ from pathlib import Path
 import iris.coords
 import iris.cube
 import matplotlib as mpl
-import matplotlib.pyplot as plt
 import numpy as np
 import pytest
 
-from CSET.operators import collapse, misc, plot, read
-
-
-def test_check_single_cube():
-    """Conversion to a single cube, and rejection where not possible."""
-    cube = iris.cube.Cube([0.0])
-    cubelist = iris.cube.CubeList([cube])
-    long_cubelist = iris.cube.CubeList([cube, cube])
-    non_cube = 1
-    assert plot._check_single_cube(cube) == cube
-    assert plot._check_single_cube(cubelist) == cube
-    with pytest.raises(TypeError):
-        plot._check_single_cube(long_cubelist)
-    with pytest.raises(TypeError):
-        plot._check_single_cube(non_cube)
-
-
-def test_load_colorbar_map():
-    """Colorbar is loaded correctly."""
-    colorbar = plot._load_colorbar_map()
-    assert isinstance(colorbar, dict)
-    # Check we can find an example definition.
-    assert colorbar["temperature_at_screen_level"] == {
-        "cmap": "RdYlBu_r",
-        "max": 323,
-        "min": 263,
-        "ymax": "auto",
-        "ymin": "auto",
-    }
-
-
-def test_load_colorbar_map_override(tmp_path):
-    """Colorbar is loaded correctly and overridden by the user definition."""
-    # Setup a user provided colorbar override.
-    user_definition = {"temperature_at_screen_level": {"max": 1000, "min": 0}}
-    user_colorbar_file = tmp_path / "colorbar.json"
-    with open(user_colorbar_file, "wt") as fp:
-        json.dump(user_definition, fp)
-
-    colorbar = plot._load_colorbar_map(user_colorbar_file)
-
-    assert isinstance(colorbar, dict)
-    # Check definition is updated.
-    assert colorbar["temperature_at_screen_level"] == {
-        "cmap": "RdYlBu_r",
-        "max": 1000,
-        "min": 0,
-        "ymax": "auto",
-        "ymin": "auto",
-    }
-    # Check we can still see unchanged definitions.
-    assert colorbar["temperature_at_screen_level_difference"] == {
-        "cmap": "bwr",
-        "max": 10,
-        "min": -10,
-    }
-
-
-def test_load_colorbar_map_override_file_not_found(tmp_path):
-    """Colorbar overridden by the user definition in non-existent file."""
-    user_colorbar_file = tmp_path / "colorbar.json"
-    colorbar = plot._load_colorbar_map(user_colorbar_file)
-    # Check it still returns the built-in one.
-    assert isinstance(colorbar, dict)
-
-
-def test_colorbar_map_levels(cube, tmp_working_dir):
-    """Colorbar definition is found for cube."""
-    cmap, levels, norm = plot._colorbar_map_levels(cube)
-    assert cmap == mpl.pyplot.get_cmap("RdYlBu_r", 51)
-    assert (levels == np.linspace(263, 323, 101)).all()
-    assert norm is None
-
-
-def test_colorbar_map_levels_xaxis(cube, tmp_working_dir):
-    """Set levels for based on xmin, xmax."""
-    cube = iris.cube.Cube(np.arange(10), long_name="zonal_wind_at_pressure_levels")
-    cmap, levels, norm = plot._colorbar_map_levels(cube, axis="x")
-    assert cmap is None
-    assert levels == [-25, 25]
-    assert norm is None
-
-
-def test_colorbar_map_levels_xaxis_default(cube, tmp_working_dir):
-    """Test for variable without xmin, xmax."""
-    cube = iris.cube.Cube(
-        np.arange(10), long_name="zonal_wind_at_pressure_levels_difference"
-    )
-    cmap, levels, norm = plot._colorbar_map_levels(cube, axis="x")
-    assert cmap is None
-    assert levels == [-20, 20]
-    assert norm is None
-
-
-def test_colorbar_map_levels_yaxis(cube, tmp_working_dir):
-    """Set levels for based on ymin, ymax."""
-    cube = iris.cube.Cube(np.arange(10), long_name="toa_upward_shortwave_flux")
-    cmap, levels, norm = plot._colorbar_map_levels(cube, axis="y")
-    assert cmap is None
-    assert levels == [0, 500]
-    assert norm is None
-
-
-def test_colorbar_map_levels_yaxis_default(cube, tmp_working_dir):
-    """Test for variable without ymin, ymax."""
-    cube = iris.cube.Cube(
-        np.arange(10), long_name="toa_upward_shortwave_flux_difference"
-    )
-    cmap, levels, norm = plot._colorbar_map_levels(cube, axis="y")
-    assert cmap is None
-    assert levels == [-100, 100]
-    assert norm is None
-
-
-def test_colorbar_map_levels_yaxis_auto(cube, tmp_working_dir):
-    """Set levels for based on ymin, ymax set to auto."""
-    cmap, levels, norm = plot._colorbar_map_levels(cube, axis="y")
-    assert cmap is None
-    assert levels is None
-    assert norm is None
-
-
-def test_colorbar_map_levels_def_on_levels(cube, tmp_working_dir):
-    """Colorbar definition that uses levels is found for cube."""
-    cube = iris.cube.Cube(
-        np.arange(10), long_name="surface_microphysical_rainfall_rate"
-    )
-    cmap, levels, norm = plot._colorbar_map_levels(cube)
-    assert levels == [0, 0.125, 0.25, 0.5, 1, 2, 4, 8, 16, 32, 64, 128, 256]
-
-
-def test_colorbar_map_levels_def_on_levels_test_visibility_in_air(
-    cube, tmp_working_dir
-):
-    """Colorbar definition that uses levels is found for cube."""
-    cube = iris.cube.Cube(np.arange(10), long_name="visibility_in_air")
-    cmap, levels, norm = plot._colorbar_map_levels(cube)
-    assert levels == [
-        0,
-        0.05,
-        0.1,
-        0.2,
-        1.0,
-        2.0,
-        5.0,
-        10.0,
-        20.0,
-        30.0,
-        50.0,
-        70.0,
-        100.0,
-    ]
-
-
-def test_colorbar_map_levels_name_fallback(cube, tmp_working_dir):
-    """Colorbar definition is found for cube after checking its other names."""
-    cube.standard_name = None
-    cmap, levels, norm = plot._colorbar_map_levels(cube)
-    assert cmap == mpl.pyplot.get_cmap("RdYlBu_r", 51)
-    assert (levels == np.linspace(263, 323, 101)).all()
-    assert norm is None
-
-
-def test_colorbar_map_levels_unknown_variable_fallback(cube, tmp_working_dir):
-    """Colorbar definition doesn't exist for cube."""
-    cube.standard_name = None
-    cube.long_name = None
-    cube.var_name = "unknown"
-    cmap, levels, norm = plot._colorbar_map_levels(cube)
-    assert cmap == mpl.pyplot.get_cmap("viridis")
-    assert levels is None
-    assert norm is None
-
-
-def test_colorbar_map_levels_pressure_level(transect_source_cube, tmp_working_dir):
-    """Pressure level specific colorbar definition is picked up."""
-    cube_250hPa = transect_source_cube.extract(iris.Constraint(pressure=250))
-    cmap, levels, norm = plot._colorbar_map_levels(cube_250hPa)
-    assert cmap == mpl.pyplot.get_cmap("RdYlBu_r", 51)
-    assert (levels == np.linspace(200, 240, 101)).all()
-    assert norm is None
-
-
-def test_colorbar_map_levels_pressure_level_yaxis(
-    transect_source_cube, tmp_working_dir
-):
-    """Pressure level specific colorbar definition is picked up."""
-    cube_250hPa = transect_source_cube.extract(iris.Constraint(pressure=250))
-    cube_250hPa.rename("zonal_wind_at_pressure_levels")
-    cmap, levels, norm = plot._colorbar_map_levels(cube_250hPa, axis="y")
-    assert cmap is None
-    assert levels == [-20, 20]
-    assert norm is None
-
-
-def test_colorbar_map_levels_missing_pressure_level(
-    transect_source_cube, tmp_working_dir, caplog
-):
-    """Pressure level specific colorbar definition is not defined."""
-    cube_288hPa = transect_source_cube.extract(iris.Constraint(pressure=250))
-    cube_288hPa.coord("pressure").points = 288.0
-    cmap, levels, norm = plot._colorbar_map_levels(cube_288hPa)
-    with caplog.at_level(logging.DEBUG):
-        cmap, levels, norm = plot._colorbar_map_levels(cube_288hPa)
-        assert caplog.record_tuples == [
-            (
-                "root",
-                logging.DEBUG,
-                "temperature_at_pressure_levels has no colorbar definition for pressure level 288.",
-            ),
-            (
-                "root",
-                logging.DEBUG,
-                "Using min and max for temperature_at_pressure_levels colorbar.",
-            ),
-        ]
+from CSET.operators import collapse, plot, read
 
 
 def test_setup_spatial_map(cube):
@@ -394,175 +178,6 @@ def test_set_title_and_filename_year_one(cube):
     )
     assert plot_filename == "recipe.png"
     assert plot_title == "recipe"
-
-
-def test_colorbar_map_mask(cube, tmp_working_dir):
-    """Test to ensure axis picks up correct colormap for a mask."""
-    cube.rename(f"mask_for_{cube.name()}")
-    cmap, levels, norm = plot._colorbar_map_levels(cube)
-    assert cmap == mpl.colors.ListedColormap(["w", "dodgerblue"])
-    assert levels == [0, 1, 2]
-    assert isinstance(norm, mpl.colors.BoundaryNorm)
-    assert (norm.boundaries == levels).all()
-
-
-def test_colorbar_map_beaufort_scale(cube, tmp_working_dir):
-    """Test to ensure picks up correct colormap for a cube in Beaufort Scale."""
-    cube.rename("wind_speed_at_10m_on_Beaufort_Scale")
-    cmap, levels, norm = plot._colorbar_map_levels(cube)
-    assert cmap == mpl.colors.ListedColormap(
-        [
-            "black",
-            (0, 0, 0.6),
-            "blue",
-            "cyan",
-            "green",
-            "yellow",
-            (1, 0.5, 0),
-            "red",
-            "pink",
-            "magenta",
-            "purple",
-            "maroon",
-            "white",
-        ]
-    )
-    assert levels == [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13]
-    assert isinstance(norm, mpl.colors.BoundaryNorm)
-    assert (norm.boundaries == levels).all()
-
-
-def test_colorbar_map_mask_difference(cube, tmp_working_dir):
-    """Test to ensure axis picks up correct colormap for a mask difference."""
-    cube.rename(f"mask_for_{cube.name()}_difference")
-    cmap, levels, norm = plot._colorbar_map_levels(cube)
-    assert cmap == mpl.colors.ListedColormap(["goldenrod", "w", "teal"])
-    assert levels == [-2, -0.5, 0.5, 2]
-    assert isinstance(norm, mpl.colors.BoundaryNorm)
-    assert (norm.boundaries == levels).all()
-
-
-def test_colorbar_map_beaufort_scale_difference(cube, tmp_working_dir):
-    """Test to ensure picks up correct colormap for Beaufort Scale difference."""
-    cube.rename("wind_speed_at_10m_on_Beaufort_Scale_difference")
-    cmap, levels, norm = plot._colorbar_map_levels(cube)
-    assert cmap == plt.get_cmap("bwr", 8)
-    assert levels == [
-        -3.5,
-        -2.5,
-        -1.5,
-        -0.5,
-        0.5,
-        1.5,
-        2.5,
-        3.5,
-    ]
-    assert isinstance(norm, mpl.colors.BoundaryNorm)
-    assert (norm.boundaries == levels).all()
-
-
-def test_colorbar_map_axis_mask(cube, tmp_working_dir):
-    """Test to ensure axis picks up correct levels when mask defined."""
-    cube.rename(f"mask_for_{cube.name()}")
-    cmap, levels, norm = plot._colorbar_map_levels(cube, axis="y")
-    assert cmap is None
-    assert levels == [0, 1]
-    assert norm is None
-
-
-def test_colorbar_map_axis_mask_difference(cube, tmp_working_dir):
-    """Test to ensure axis picks up correct levels when mask difference defined."""
-    cube.rename(f"mask_for_{cube.name()}_difference")
-    cmap, levels, norm = plot._colorbar_map_levels(cube, axis="x")
-    assert cmap is None
-    assert levels == [-1, 1]
-    assert norm is None
-
-
-def test_colorbar_map_beaufort_scale_axis(cube, tmp_working_dir):
-    """Test to ensure axis picks up correct levels for a cube in Beaufort Scale."""
-    cube.rename("wind_speed_at_10m_on_Beaufort_Scale")
-    cmap, levels, norm = plot._colorbar_map_levels(cube, axis="y")
-    assert cmap is None
-    assert levels == [0, 12]
-    assert norm is None
-
-
-def test_colorbar_map_beaufort_scale_axis_difference(cube, tmp_working_dir):
-    """Test to ensure axis picks up correct levels for a cube in Beaufort Scale difference."""
-    cube.rename("wind_speed_at_10m_on_Beaufort_Scale_difference")
-    cmap, levels, norm = plot._colorbar_map_levels(cube, axis="x")
-    assert cmap is None
-    assert levels == [-4, 4]
-    assert norm is None
-
-
-def test_colorbar_map_celsius(cube, tmp_working_dir):
-    """Test to ensure color bar is changed for temperature in Celsius."""
-    cube = misc.convert_units(cube, "Celsius")
-    cmap = mpl.cm.RdYlBu
-    norm = None
-    levels = [273, 373]
-    cmap, levels, norm = plot._custom_colormap_celsius(
-        cube, cmap=cmap, levels=levels, norm=norm
-    )
-    assert cmap == mpl.cm.RdYlBu
-    assert norm is None
-    assert levels == [0, 100]
-
-
-def test_colorbar_map_probabilities_axis(cube, tmp_working_dir):
-    """Test to ensure axis picks up correct levels for a cube of probabilities."""
-    cube.rename("probability_of_temperature_>_276")
-    cmap, levels, norm = plot._colorbar_map_levels(cube, axis="x")
-    assert cmap is None
-    assert levels == [0, 1]
-    assert norm is None
-
-
-def test_colorbar_map_probabilities(cube, tmp_working_dir):
-    """Test to ensure colorbar picks up correct maap for cube of probabilities."""
-    cube.rename("probability_of_temperature_>_276")
-    cmap, levels, norm = plot._colorbar_map_levels(cube)
-    assert cmap == mpl.colors.ListedColormap(
-        [
-            "#FFFFFF",
-            "#636363",
-            "#e1dada",
-            "#B5CAFF",
-            "#8FB3FF",
-            "#7F97FF",
-            "#ABCF63",
-            "#E8F59E",
-            "#FFFA14",
-            "#FFD121",
-            "#FFA30A",
-        ]
-    )
-    assert levels == [0.0, 0.01, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0]
-    assert isinstance(norm, mpl.colors.BoundaryNorm)
-    assert (norm.boundaries == levels).all()
-
-
-def test_colorbar_map_aviation_colour_state(cube, tmp_working_dir):
-    """Test to ensure color bar is change for aviation colour states."""
-    cube.rename("aviation_colour_state")
-    expected_levels = [-0.5, 0.5, 1.5, 2.5, 3.5, 4.5, 5.5, 6.5]
-    expected_colors = [
-        "#87ceeb",
-        "#ffffff",
-        "#8ced69",
-        "#ffff00",
-        "#ffd700",
-        "#ffa500",
-        "#fe3620",
-    ]
-    expected_cmap = mpl.colors.ListedColormap(expected_colors)
-    cmap, levels, norm = plot._custom_colormap_aviation_colour_state(cube)
-    assert cmap == expected_cmap
-    assert levels == expected_levels
-    assert isinstance(norm, mpl.colors.BoundaryNorm)
-    assert (norm.boundaries == levels).all()
 
 
 def test_spatial_contour_plot(cube, tmp_working_dir):
@@ -1043,49 +658,6 @@ def test_plot_and_save_postage_stamps_in_single_plot_power_spectrum_series(
         histtype="step",
     )
     assert Path("test.png").is_file()
-
-
-def test_create_alpha_matrix_shape():
-    """Test shape of alpha matrix used in power spectrum calculation."""
-    Ny, Nx = 10, 15
-    alpha = plot._create_alpha_matrix(Ny, Nx)
-    assert alpha.shape == (Ny, Nx), "Alpha matrix shape mismatch"
-
-
-def test_create_alpha_matrix_values():
-    """Test alpha matrix contains only positive values."""
-    Ny, Nx = 4, 4
-    alpha = plot._create_alpha_matrix(Ny, Nx)
-    assert np.all(alpha >= 0), "Alpha matrix contains negative values"
-    assert np.isclose(alpha[0, 0], np.sqrt((1 / Nx) ** 2 + (1 / Ny) ** 2)), (
-        "Alpha matrix value incorrect"
-    )
-
-
-def test_dct_ps_output_shape():
-    """Test shape of power spectrum output from _DCT_ps."""
-    Nt, Ny, Nx = 5, 10, 10
-    y_3d = np.random.rand(Nt, Ny, Nx)
-    ps = plot._DCT_ps(y_3d)
-    expected_shape = (Nt, min(Nx - 1, Ny - 1))
-    assert ps.shape == expected_shape, "Power spectrum output shape mismatch"
-
-
-def test_dct_ps_non_negative():
-    """Test power spectrum only contains positive values."""
-    Nt, Ny, Nx = 3, 8, 8
-    y_3d = np.random.rand(Nt, Ny, Nx)
-    ps = plot._DCT_ps(y_3d)
-    assert np.all(ps >= 0), "Power spectrum contains negative values"
-
-
-def test_dct_ps_known_input():
-    """Test _DCT_ps produces non-zero spectrum for constant input."""
-    # Use a constant field to test expected behavior
-    Nt, Ny, Nx = 2, 4, 4
-    y_3d = np.ones((Nt, Ny, Nx))
-    ps = plot._DCT_ps(y_3d)
-    assert np.allclose(ps[:, 1:], 0, atol=1e-6), "Non-zero spectrum for constant input"
 
 
 def test_plot_power_spectrum_no_sequence_coordinate(field2d_cube, tmp_working_dir):
